backend/requirements.txt
# Simple FastAPI Backend for Meter Reading
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML - Minimal dependencies
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0
Pillow==10.0.1
numpy==1.24.3